+++ plugins/org.eclipse.osee.framework.access/src/org/eclipse/osee/framework/access/AccessObject.java
@@ -24,7 +24,7 @@ public abstract class AccessObject {
 
    public abstract void removeFromDatabase(int subjectId) throws OseeCoreException;
 
-   public abstract int getId();
+   public abstract long getId();
 
    public static AccessObject getAccessObject(Object object) throws OseeCoreException {
       if (object instanceof Artifact) {
